+++ build-docs/Iteration2.md
@@ -2,6 +2,8 @@
 
 ### clean process
 * separate in util and process -> 20/10/2023 30 mins
-* create test for each process function
-* create comment of each function and process
+* create test for each process function -> 06/12/2023 30 mins
+* create comment of each function and process -> 06/12/2023 10 mins (I don't)
+{ https://pkg.go.dev/golang.org/x/tools/cmd/godoc#section-sourcefiles }
+Install extension minlify doc writer
 * clean args and load all from file runner.pmsgo in the same folder than the binary
