+++ build-docs/Iteration3.md
@@ -1,6 +1,6 @@
 ## Iteration 3
 
 ### logs and test
-
+* Create global workdir and run files with relative path to this
 * create a sqlite log rotate with max size log file and access from global channel
 * test for everything
