+++ go.mod
@@ -1,3 +1,9 @@
 module github.com/dobycontech/pmsgo
 
 go 1.21.3
+
+require (
+	github.com/yuin/goldmark v1.4.13 // indirect
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/tools v0.16.0 // indirect
+)
