+++ test/ps.json
@@ -1,5 +1,5 @@
 {
     "name": "psname1",
-    "bin":"./test/psname1/psname1",
+    "bin":"../test/psname1/psname1",
     "timeoutseconds":-1
 }
