+++ test/psname1/main.go
@@ -1,9 +1,10 @@
+// Package main: The main function waits for 5 minutes before printing "end of process".
 package main
 
 import "time"
 
-func main(){
+func main() {
 	println("start of process")
-	<-time.After(5*time.Minute)
+	<-time.After(5 * time.Minute)
 	println("end of process")
 }
